backend/recommendation engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
langchain-community==0.2.4
langchain-chroma==0.1.2
langchain-text-splitters==0.2.2
sentence-transformers==3.0.1
pandas==2.2.2
numpy==1.26.4
redis==5.0.6
python-dotenv==1.0.1
chromadb==0.5.3
